+++ src/Symfony/Component/Form/Extension/Core/Type/MoneyType.php
@@ -34,7 +34,8 @@ class MoneyType extends AbstractType
                 $options['grouping'],
                 $options['rounding_mode'],
                 $options['divisor'],
-                $options['html5'] ? 'en' : null
+                $options['html5'] ? 'en' : null,
+                $options['model_type'],
             ))
         ;
     }
@@ -59,6 +60,7 @@ class MoneyType extends AbstractType
             'compound' => false,
             'html5' => false,
             'invalid_message' => 'Please enter a valid money amount.',
+            'model_type' => 'float',
         ]);
 
         $resolver->setAllowedValues('rounding_mode', [
@@ -75,9 +77,19 @@ class MoneyType extends AbstractType
 
         $resolver->setAllowedTypes('html5', 'bool');
 
+        $resolver->setAllowedValues('model_type', ['float', 'integer']);
+
         $resolver->setNormalizer('grouping', static function (Options $options, $value) {
             if ($value && $options['html5']) {
                 throw new LogicException('Cannot use the "grouping" option when the "html5" option is enabled.');
+            }
+
+            return $value;
+        });
+
+        $resolver->setNormalizer('model_type', static function (Options $options, $value) {
+            if ('integer' === $value && 1 === $options['divisor']) {
+                throw new LogicException('When the "model_type" option is set to "integer", the "divisor" option should not be set to "1".');
             }
 
             return $value;