+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Deprecate not configuring the `default_protocol` option of the `UrlType`, it will default to `null` in 8.0
  * Add a `keep_as_list` option to `CollectionType`
+ * Add a new `model_type` option to `MoneyType`, to be able to cast the transformed value to `integer`
 
 7.0
 ---